backend/requirements.txt
aioredis==1.3.1
APScheduler==3.6.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
coverage==6.4.2
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.4
django-extensions==3.1.3
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-filters==1.0.0.dev2
djangorestframework-simplejwt==4.7.0
factory_boy==3.2.1
hiredis==2.0.0
honcho==1.0.1
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
msgpack==1.0.2
Pillow==8.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyserial==3.5
python-telegram-bot==13.5
pytz==2021.1
requests==2.21.0
service-identity==21.1.0
shamrock==0.2.0
six==1.16.0
sqlparse==0.4.1
tornado==6.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.24.3
zope.interface==5.4.0
